requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudsight==0.9.2
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
mediapipe==0.9.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==9.0.0
MouseInfo==0.1.3
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.23.5
openai==0.26.5
openai-whisper==20230124
opencv-contrib-python==4.7.0.68
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
PyAudio==0.2.13
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
semantic-version==2.10.0
setuptools==65.6.3
setuptools-rust==1.5.2
six==1.16.0
stack-data==0.6.2
tokenizers==0.13.2
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
wheel==0.37.1
wincertstore==0.2
yarg==0.1.9
yarl==1.8.2
zipp==3.14.0
